Made CppCheck happy (with 1 exception (false positive)) Might've done some other things, I think it's been a while since I last commited.
git-svn-id: https://jarenaissance.svn.sourceforge.net/svnroot/jarenaissance/trunk@55 7d4315b7-3a6b-433c-8ee3-f64c52bfdc48

+++ code/jarlib/include/jar/core/Time.hpp
@@ -33,8 +33,8 @@ const float TimeTypesPerSecond = 1000;
 /// Conversion from TimeType to seconds
 const float SecondsPerTimeType = 1.0f / TimeTypesPerSecond;
 
-/// The type of time variables
-typedef long int TimeType; //when you change this, check GetTime() for validity!
+/// The type of time variables - unsigned int, will overflow every 49 days
+typedef unsigned int TimeType; //when you change this, check GetTime() for validity!
 
 JARCOREAPI TimeType GetTime();
